requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
flask-bcrypt==1.0.1
pymongo==4.5.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
Pillow==10.2.0
